beerview/views.py

from rest_framework.generics import ListCreateAPIView, RetrieveUpdateAPIView, RetrieveUpdateDestroyAPIView
from rest_framework.permissions import IsAuthenticatedOrReadOnly

from .models import Beer
from .serializers import PostSerializer

class PostsList(ListCreateAPIView):
    queryset = Beer.objects.all()
    serializer_class = PostSerializer

class PostsDetail(RetrieveUpdateDestroyAPIView):
    permission_classes = (IsAuthenticatedOrReadOnly,)
    queryset = Beer.objects.all()
    serializer_class = PostSerializer
